Move subtask_accuracy metric hooks to to agent
